requirements.txt
autopep8==1.5.2
certifi==2020.6.20
chardet==3.0.4
Django==2.2.3
django-admin-thumbnails==0.2.5
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
idna==2.10
Pillow==7.2.0
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stripe==2.49.0
urllib3==1.25.9
